+++ NEWS.md
@@ -2,6 +2,7 @@
 
 -   Add ability to specify column order with matrix_table function
 -   Set default order for multi_table for results to be shown from high to low remove frequency summary row for non-grouped data
+-   Bug fix in matrix_mean that caused standard errors to be underestimated for weighted data
 
 # surveyexplorer 0.1.0